+++ psdist/visualization/visualization.py
@@ -7,14 +7,12 @@ from matplotlib import pyplot as plt
 import numpy as np
 import proplot as pplt
 import scipy.optimize
-import scipy.stats
 
-from ..image import project as project_image
-from ..utils import random_selection
-from ..utils import centers_from_edges
-
-from . import discrete as _vis_discrete
-from . import image as _vis_image
+import psdist.image
+import psdist.discrete
+import psdist.utils as utils
+import psdist.visualization.discrete as psv_discrete
+import psdist.visualization.image as psv_image
 
 
 # General
@@ -91,62 +89,51 @@ def linear_fit(x, y):
     return yfit, slope, intercept
 
 
-def plot1d(x, y, ax=None, offset=0.0, flipxy=False, kind='line', **kws):
-    """Convenience function for one-dimensional line/step/bar plots."""  
-    y = y.copy()
+def plot1d(x, y, ax=None, offset=0.0, flipxy=False, kind="line", **kws):
+    """Convenience function for one-dimensional line/step/bar plots."""
     func = ax.plot
-    if kind in ['line', 'step']:
+    if kind in ["line", "step"]:
         if flipxy:
             func = ax.plotx
         else:
             func = ax.plot
-        if kind == 'step':
-            kws.setdefault('drawstyle', 'steps-mid')
-    elif kind in ['linefilled', 'stepfilled']:
+        if kind == "step":
+            kws.setdefault("drawstyle", "steps-mid")
+    elif kind in ["linefilled", "stepfilled"]:
         if flipxy:
             func = ax.fill_betweenx
         else:
             func = ax.fill_between
-        kws.setdefault('alpha', 1.0)
-        if kind == 'stepfilled':
-            kws.setdefault('step', 'mid')
-    elif kind == 'bar':
+        kws.setdefault("alpha", 1.0)
+        if kind == "stepfilled":
+            kws.setdefault("step", "mid")
+    elif kind == "bar":
         if flipxy:
             func = ax.barh
         else:
             func = ax.bar
-            
+
     # Handle offset
-    if kind == 'bar':
-        kws['left' if flipxy else 'bottom'] = offset
+    if kind == "bar":
+        kws["left" if flipxy else "bottom"] = offset * np.ones(len(x))
         return func(x, y, **kws)
-    elif kind in ['linefilled', 'stepfilled']:
+    elif kind in ["linefilled", "stepfilled"]:
         return func(x, offset, y + offset, **kws)
     return func(x, y + offset, **kws)
 
 
-# Image
-# ------------------------------------------------------------------------------
-# Corner plot
+# The following functions set up a corner plot. It's probably better
+# to have some object-oriented approach like in Seaborn?
 
-def _set_corner_limits(axes, limits, diag=False):
-    for i in range(axes.shape[1]):
-        axes[:, i].format(xlim=limits[i])
-    start = int(diag)
-    for i, lim in enumerate(limits[1:], start=start):
-        axes[i, :(i + 1 - start)].format(ylim=lim)
-    return axes
-    
 
-def _setup_corner(n, diag, labels, limits=None, **fig_kws):
-    """Set up corner plot axes."""
+def _corner_make_figure(n, diag, labels, limits=None, **fig_kws):
     if labels is None:
         labels = n * [""]
     nrows = ncols = n if diag else n - 1
     start = 1 if diag else 0
     fig_kws.setdefault("figwidth", 1.5 * nrows)
     fig_kws.setdefault("aligny", True)
-    fig, axes = pplt.subplots(
+    fig, axs = pplt.subplots(
         nrows=nrows,
         ncols=ncols,
         sharex=1,
@@ -158,254 +145,234 @@ def _setup_corner(n, diag, labels, limits=None, **fig_kws):
     for i in range(nrows):
         for j in range(ncols):
             if j > i:
-                axes[i, j].axis("off")
-    for ax, label in zip(axes[-1, :], labels):
+                axs[i, j].axis("off")
+    for ax, label in zip(axs[-1, :], labels):
         ax.format(xlabel=label)
-    for ax, label in zip(axes[(1 if diag else 0) :, 0], labels[1:]):
+    for ax, label in zip(axs[(1 if diag else 0) :, 0], labels[1:]):
         ax.format(ylabel=label)
     for i in range(nrows):
-        axes[:-1, i].format(xticklabels=[])
-        axes[i, 1:].format(yticklabels=[])
-    for ax in axes:
+        axs[:-1, i].format(xticklabels=[])
+        axs[i, 1:].format(yticklabels=[])
+    for ax in axs:
         ax.format(xspineloc="bottom", yspineloc="left")
     if diag:
         for i in range(n):
-            axes[i, i].format(yspineloc="neither")
+            axs[i, i].format(yspineloc="neither")
     if limits is not None:
-        _set_corner_limits(axes, limits, diag=diag)
-    axes.format(
+        _set_corner_limits(axs, limits, diag=diag)
+    axs.format(
         xtickminor=True, ytickminor=True, xlocator=("maxn", 3), ylocator=("maxn", 3)
     )
-    return fig, axes
+    return fig, axs
 
 
-def corner(
-    data,
-    kind="hist",
-    diag_kind="step",
-    coords=None,
-    limits=None,
-    labels=None,
-    samples=None,
-    diag_height_frac=0.6,
-    autolim_kws=None,
-    diag_kws=None,
-    fig_kws=None,
-    prof=False,
-    prof_kws=None,
-    return_fig=False,
-    return_mesh=False,
-    axes=None,
-    **plot_kws,
-):
-    """Plot one- and two-dimensional projections in a corner plot.
+def _corner_set_limits(axs, limits, diag=False):
+    for i in range(axs.shape[1]):
+        axs[:, i].format(xlim=limits[i])
+    start = int(diag)
+    for i, lim in enumerate(limits[1:], start=start):
+        axs[i, : (i + 1 - start)].format(ylim=lim)
+    return axs
 
-    Parameters
-    ----------
-    data : ndarray
-        If `data.ndim == 2`, we have the coordinates of k points in n-dimensional
-        space; otherwise, we have an n-dimensional image.
-    kind : {'hist', 'scatter'}
-        The kind of 2D plot to make if `data` is a list of points.
-            'hist': calls `hist`
-            'scatter': `scatter`
-            'kde': calls `kde`
-    diag_kind : {'line', 'step', 'bar', 'None'}
-        Kind of 1D plot on diagonal axes. Any variant of 'None', 'none', None
-        will remove the diagonal axes from the figure, resulting in a D-1 x D-1
-        array of subplots instead of a D x D array.
-    coords : list[ndarray]
-        Coordinates along each axis of the grid (if `data` is an image).
-    limits : list[tuple]
-        The (min, max) coordinates for each dimension. This is used to set the
-        axis limits, as well as for data binning if plotting a histogram.
-    labels : list[str]
-        The axis labels.
-    samples : int or float
-        Number of samples to use in bivariate plots. If less than 1, specifies
-        the fraction of points. All one-dimensional histograms are computed
-        before downsampling.
-    diag_height_frac : float
-        Reduce the height of 1D profiles (diagonal subplots) relative to the
-        y axis height.
-    autolim_kws : dict
-        Key word arguments passed to `autolim`.
-    diag_kws : dict
-        Key word arguments passed to 1D plotting function on diagonal axes.
-    fig_kws : dict
-        Key word arguments passed to `pplt.subplots` such as 'figwidth'.
-    prof : bool or 'edges'
-        Whether to overlay 1D profiles on 2D plots in off-diagonal axes. If
-        'edges', only plot profiles in the left column and bottom row of
-        the figure. This is a good option if not using diagonal subplots.
-    prof_kws : dict
-        Key word arguments passed to `image_profiless`.
-    return_fig : bool
-        Whether to return `fig` in addition to `axes`.
-    return_mesh : bool
-        Whether to also return a mesh from one of the pcolor plots. This is
-        useful if you want to put a colorbar on the figure later.
-    axes : proplot.gridspec
-        Plot on an existing figure.
-    **plot_kws
-        Key word arguments passed to 2D plotting function.
 
-    Returns
-    -------
-    axes : proplot.gridspec
-        Array of subplot axes.
-    Optional:
-        fig : proplot.figure
-            Proplot figure object.
-        mesh : matplotlib.collections.QuadMesh
-            Mesh from the latest call of `pcolormesh`. This is helpful if you want a
-            global colorbar.
-    """
-    # Determine whether data is point cloud or image.
-    n = data.ndim
-    pts = False
-    if n == 2:
-        n = data.shape[1]
-        pts = True
+class CornerGrid:
+    def __init__(
+        self, 
+        n=4, 
+        diag=True, 
+        labels=None, 
+        limits=None, 
+        diag_height_frac=0.6, 
+        **fig_kws
+    ):
+        # Create figure.
+        self.n = n
+        self.diag = diag
+        self.labels = labels
+        self.limits = limits
+        self.diag_height_frac = 0.6
+        self.start = int(self.diag)
+        if self.labels is None:
+            self.labels = self.n * [""]
+        self.nrows = self.ncols = self.n
+        if not self.diag:
+            self.nrows = self.nrows - 1
+            self.ncols = self.ncols - 1
+        self.fig_kws = fig_kws
+        self.fig_kws.setdefault("figwidth", 1.5 * self.nrows)
+        self.fig_kws.setdefault("aligny", True)
+        self.fig, self.axs = pplt.subplots(
+            nrows=self.nrows,
+            ncols=self.ncols,
+            sharex=1,
+            sharey=1,
+            spanx=False,
+            spany=False,
+            **self.fig_kws,
+        )
 
-    # Parse arguments
-    diag = diag_kind in ["line", "bar", "step", "linefilled", "stepfilled"]
-    start = 1 if diag else 0
-    if diag_kws is None:
-        diag_kws = dict()
-    diag_kws.setdefault("color", "black")
-    diag_kws.setdefault("lw", 1.0)
-    if pts:
-        if kind == "scatter":
-            plot_kws.setdefault("s", 6)
-            plot_kws.setdefault("c", "black")
-            plot_kws.setdefault("marker", ".")
-            plot_kws.setdefault("ec", "none")
-            if "color" in plot_kws:
-                plot_kws["c"] = plot_kws.pop("color")
-            if "ms" in plot_kws:
-                plot_kws["s"] = plot_kws.pop("ms")
-            if "cmap" in plot_kws:
-                plot_kws.pop("cmap")
-        elif kind == "hist":
-            plot_kws.setdefault("ec", "None")
-            plot_kws.setdefault("process_kws", dict(mask_zero=True))
-    else:
-        plot_kws.setdefault("ec", "None")
-
-    # Compute ax limits.
-    if (not pts) and (coords is None):
-        coords = [np.arange(s) for s in data.shape]
-    if autolim_kws is None:
-        autolim_kws = dict()
-    if limits is None:
-        if pts:
-            limits = _vis_discrete.auto_limits(data, **autolim_kws)
+        # Collect diagonal/off-diagonal subplots.
+        self.axs_diag, self.axs_offdiag = [], []
+        self.axis_diag, self.axis_offdiag = [], []
+        if self.diag:
+            for i in range(self.n):
+                self.axs_diag.append(self.axs[i, i])
+                self.axis_diag.append(i)
+            for i in range(1, self.n):
+                for j in range(i):
+                    self.axs_offdiag.append(self.axs[i, j])
+                    self.axis_offdiag.append((j, i))
         else:
-            limits = [(c[0], c[-1]) for c in coords]
+            for i in range(self.n - 1):
+                for j in range(i + 1):
+                    self.axs_offdiag.append(self.axs[i, j])
+                    self.axis_offdiag.append((j, i + 1))
 
-    # Create the figure.
-    if axes is None:
-        if fig_kws is None:
-            fig_kws = dict()
-        fig, axes = _setup_corner(n, diag, labels, limits, **fig_kws)
-    else:
-        # We are plotting on an existing figure. Expand the existing ax limits
-        # based on our new data.
-        if axes.shape[1] == n:
-            old_limits = [axes[i, i].get_xlim() for i in range(n)]
+        # 
+        for i in range(self.nrows):
+            for j in range(self.ncols):
+                if j > i:
+                    self.axs[i, j].axis('off')
+        for ax, label in zip(self.axs[-1, :], self.labels):
+            ax.format(xlabel=label)
+        for ax, label in zip(self.axs[self.start :, 0], self.labels[1:]):
+            ax.format(ylabel=label)
+        for i in range(self.nrows):
+            self.axs[:-1, i].format(xticklabels=[])
+            self.axs[i, 1:].format(yticklabels=[])
+        for ax in self.axs:
+            ax.format(xspineloc='bottom', yspineloc='left')
+        for ax in self.axs_diag:
+            ax.format(yspineloc='neither')
+        self.axs.format(xtickminor=True, ytickminor=True, 
+                        xlocator=('maxn', 3), ylocator=('maxn', 3))
+        self.set_limits(limits)
+
+    def get_limits(self):
+        if self.diag:
+            limits = [ax.get_xlim() for ax in self.axs_diag]
         else:
-            old_limits = [axes[-1, i].get_xlim() for i in range(n - 1)] + [axes[-1, 0].get_ylim()]
-        limits = [
-            (min(old_lim[0], lim[0]), max(old_lim[1], lim[1]))
-            for old_lim, lim in zip(old_limits, limits)
-        ]
-        _set_corner_limits(axes, limits, diag=(axes.shape[1] == n))
-        # Don't return fig because we do not have access to it.
-        return_fig = False
+            limits = [self.axs[-1, i].get_xlim() for i in range(self.n - 1)]
+            limits = limits + [self.axs[-1, 0].get_ylim()]
+        return limits
 
-    # Discrete points
-    if pts:
-        # Univariate plots
-        bins = "auto"
-        if "bins" in plot_kws:
-            bins = plot_kws.pop("bins")
-        edges, centers = [], []
-        for i in range(n):
-            heights, _edges = np.histogram(data[:, i], bins, limits[i], density=True)
-            heights = heights / np.max(heights)
-            _centers = centers_from_edges(_edges)
+    def set_limits(self, limits=None, expand=False):
+        if limits is not None:
+            if expand:
+                limits = np.array(limits)
+                limits_old = np.array(self.get_limits())
+                mins = np.minimum(limits[:, 0], limits_old[:, 0])
+                maxs = np.maximum(limits[:, 1], limits_old[:, 1])
+                limits = list(zip(mins, maxs))
+            for i in range(self.axs.shape[1]):
+                self.axs[:, i].format(xlim=limits[i])
+            for i, lim in enumerate(limits[1:], start=self.start):
+                self.axs[i, : (i + 1 - self.start)].format(ylim=lim)
+        self.limits = self.get_limits()
+
+    def scale_diag_ylim(self, factor):
+        for ax in self.axs_diag:
+            ymin, ymax = ax.get_ylim()
+            ax.format(ylim=(0.0, factor * ymax))
+
+    def plot_image(
+        self,
+        f,
+        coords=None,
+        prof_edge_only=False,
+        update_limits=True,
+        diag_kws=None,
+        **kws
+    ):
+        if diag_kws is None:
+            diag_kws = dict()
+        diag_kws.setdefault('color', 'black')
+        diag_kws.setdefault('lw', 1.0)
+        diag_kws.setdefault('kind', 'step')
+        kws.setdefault('kind', 'pcolor')
+        kws.setdefault('profx', False)
+        kws.setdefault('profy', False)
+
+        if coords is None:
+            coords = [np.arange(f.shape[i]) for i in range(f.ndim)]
+
+        if update_limits:
+            edges = [utils.edges_from_centers(c) for c in coords]
+            limits = [(np.min(e), np.max(e)) for e in edges]
+            self.set_limits(limits, expand=update_limits)
+
+        # Univariate plots.
+        for ax, axis in zip(self.axs_diag, self.axis_diag):
+            profile = psdist.image.project(f, axis=axis)
+            profile = profile / np.max(profile)
+            plot1d(coords[axis], profile, ax=ax, **diag_kws)
+
+        # Bivariate plots.
+        profx, profy = [kws.pop(key) for key in ('profx', 'profy')]
+        for ax, axis in zip(self.axs_offdiag, self.axis_offdiag):
+            if prof_edge_only:
+                if profx: 
+                    kws['profx'] = axis[1] == self.n - 1
+                if profy:
+                    kws['profy'] = axis[0] == 0
+            _f = psdist.image.project(f, axis=axis)
+            _f = _f / np.max(_f)
+            _coords = [coords[i] for i in axis]
+            psv_image.plot2d(_f, coords=_coords, ax=ax, **kws)
+
+        # Modify diagonal y axis limits.
+        if self.diag_height_frac is not None:
+            self.scale_diag_ylim(1.0 / self.diag_height_frac)
+
+    def plot_discrete(
+        self,
+        X,
+        diag_kws=None,
+        limits=None,
+        autolim_kws=None,
+        prof_edge_only=False,
+        update_limits=True,
+        **kws
+    ):
+        n = X.shape[1]
+        if diag_kws is None:
+            diag_kws = dict()
+        diag_kws.setdefault('color', 'black')
+        diag_kws.setdefault('lw', 1.0)
+        diag_kws.setdefault('kind', 'step')
+        kws.setdefault('kind', 'hist')
+        kws.setdefault('profx', False)
+        kws.setdefault('profy', False)
+
+        if limits is None:
+            if autolim_kws is None:
+                autolim_kws = dict()
+            limits = psv_discrete.auto_limits(X, **autolim_kws)
+        self.set_limits(limits, expand=update_limits)
+
+        # Univariate plots. Remember histogram bins and use them for 2D histograms.
+        bins = 'auto'
+        if 'bins' in kws:
+            bins = kws.pop('bins')
+        edges = []
+        for ax, axis in zip(self.axs_diag, self.axis_diag):
+            heights, _edges = np.histogram(X[:, axis], bins, limits[axis], density=True)
             edges.append(_edges)
-            centers.append(_centers)
-            if diag:
-                plot1d(_centers, heights, ax=axes[i, i], kind=diag_kind, **diag_kws)
+            plot1d(
+                utils.centers_from_edges(_edges), 
+                heights / np.max(heights), 
+                ax=ax, 
+                **diag_kws
+            )
 
-        # Take random sample of points.
-        idx = random_selection(np.arange(data.shape[0]), samples)
-
-        # Bivariate plots
-        for ii, i in enumerate(range(start, axes.shape[0])):
-            for j in range(ii + 1):
-                ax = axes[i, j]
-                if kind == "scatter":
-                    ax.scatter(data[idx, j], data[idx, ii + 1], **plot_kws)
-                elif kind == "hist":
-                    _im, _, _ = np.histogram2d(
-                        data[:, j], data[:, ii + 1], (edges[j], edges[ii + 1])
-                    )
-                    if prof == "edges":
-                        profy = j == 0
-                        profx = i == axes.shape[0] - 1
-                    else:
-                        profx = profy = prof
-                    ax = _vis_image.plot2d(
-                        _im,
-                        coords=(centers[j], centers[ii + 1]),
-                        ax=ax,
-                        kind='pcolor',
-                        profx=profx,
-                        profy=profy,
-                        prof_kws=prof_kws,
-                        **plot_kws,
-                    )
-    # Multidimensional image
-    else:
-        # Bivariate plots
-        for ii, i in enumerate(range(start, axes.shape[0])):
-            for j in range(ii + 1):
-                ax = axes[i, j]
-                if prof == 'edges':
-                    profy = j == 0
-                    profx = i == axes.shape[0] - 1
-                else:
-                    profx = profy = prof
-                _image = project_image(data, (j, ii + 1))
-                _image = _image / np.max(_image)
-                ax, mesh = _vis_image.plot2d(
-                    _image, coords=(coords[j], coords[ii + 1]),
-                    ax=ax,
-                    kind='pcolor',
-                    profx=profx,
-                    profy=profy,
-                    prof_kws=prof_kws,
-                    return_mesh=True,
-                    **plot_kws,
-                )
-        # Univariate plots
-        if diag:
-            for i in range(n):
-                profile = project_image(data, i)
-                profile = profile / np.max(profile)
-                if 'process_kws' in plot_kws and 'fill_value' in plot_kws['process_kws']:
-                    profile = np.ma.filled(profile, fill_value=plot_kws['fill_value'])
-                plot1d(coords[i], profile, ax=axes[i, i], kind=diag_kind, **diag_kws)
-    # Modify diagonal y axis limits.
-    if diag:
-        for i in range(n):
-            axes[i, i].set_ylim(0, 1.0 / diag_height_frac)
-    # Return items
-    if return_fig:
-        if return_mesh:
-            return fig, axes, mesh
-        return fig, axes
-    return axes
+        # Bivariate plots:
+        profx, profy = [kws.pop(key) for key in ('profx', 'profy')]
+        for ax, axis in zip(self.axs_offdiag, self.axis_offdiag):
+            if prof_edge_only:
+                if profx: 
+                    kws['profx'] = axis[1] == self.n - 1
+                if profy:
+                    kws['profy'] = axis[0] == 0
+            if kws['kind'] == 'hist':
+                kws['bins'] = [edges[axis[0]], edges[axis[1]]]
+            psv_discrete.plot2d(X[:, axis], ax=ax, **kws)